Move smoke tests into `tests` module
